requirements.txt
asgiref==3.11.0
Django==5.2.10
djangorestframework==3.16.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
